+++ src/JsonStaticClassCodec.php
@@ -67,10 +67,14 @@ class JsonStaticClassCodec implements JsonClassCodec {
 	 * ::toJsonArray() / input to ::newFromJsonArray.  If a class name is
 	 * returned here and it matches the runtime type of the value of that
 	 * array key, then type information will be omitted from the generated
-	 * JSON which can save space.
+	 * JSON which can save space.  The class name can be suffixed with `[]`
+	 * to indicate an array or list containing objects of the given class
+	 * name.
+	 *
 	 * @param class-string<T> $className
 	 * @param string $keyName
-	 * @return ?class-string
+	 * @return class-string|string|null A class string, a class string suffixed
+	 *   with `[]`, or null
 	 */
 	public function jsonClassHintFor( string $className, string $keyName ): ?string {
 		// Proxy to a static method on the class.